Made generated cmdlets use InternalExport attribute. Cleaned up commented code.
