+++ extensions/powershell/generators/psm1.internal.ts
@@ -24,15 +24,6 @@ export async function generatePsm1Internal(service: Host, project: Project) {
 
   # Export nothing to clear implicit exports
   Export-ModuleMember`);
-  // psm1.append('LoadScripts', 'Export-ScriptCmdlet -ScriptFolder $PSScriptRoot');
-  // psm1.append('LoadScripts', 'Invoke-Expression -Command (Export-ScriptCmdlet -ScriptFolder $PSScriptRoot)');
-  // psm1.append('LoadScripts', `
-  // $thing = Export-ScriptCmdlet -ScriptFolder $PSScriptRoot
-  // Write-Host $thing
-  // Invoke-Expression -Command $thing
-  // #Invoke-Expression -Command (Export-ScriptCmdlet -ScriptFolder $PSScriptRoot)
-  // `);
-  // psm1.append('LoadScripts', `Export-ScriptCmdlet -ScriptFolder $PSScriptRoot -EngineIntrinsics $ExecutionContext`);
   psm1.append('LoadScripts', `
   Get-ChildItem -Path $PSScriptRoot -Recurse -Filter '*.ps1' -File | ForEach-Object { . $_.FullName }
   Export-ModuleMember -Function (Get-ScriptCmdlet -ScriptFolder $PSScriptRoot)`);
